Testing seller routes
